# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Гармония (Harmony) API
  version: 1.0.0
  summary: REST and realtime API for the Harmony social network.
  description: |
    # Гармония API

    A social network API. All responses share one envelope so a client has a
    single code path for success and failure:

    ```json
    { "ok": true,  "data": { }, "meta": { "count": 15, "has_more": true, "next_cursor": "eyJ..." } }
    { "ok": false, "error": { "code": "not_found", "message": "Пост не найден.", "request_id": "..." } }
    ```

    ## Authentication

    After `POST /auth/login` the API issues two cookies:

    | Cookie | HttpOnly | Purpose |
    |---|---|---|
    | `harmony_access` | yes | Access token (JWT) |
    | `harmony_refresh` | yes | Refresh token, scoped to `/api/v1/auth` |
    | `harmony_csrf` | **no** | CSRF token for the double-submit pattern |

    The access cookie is `HttpOnly` so that an XSS payload cannot read it.
    The CSRF cookie is deliberately readable, because the client must echo it.

    **Every unsafe method** (`POST`, `PATCH`, `PUT`, `DELETE`) must send the CSRF
    token in the `X-CSRF-Token` header, matching the cookie. A correct `Origin`
    or `Referer` is also accepted, which keeps non-browser clients working.

    Non-browser clients may instead send `Authorization: Bearer <access_token>`.

    ## Resources

    Resources are addressed by an opaque 32-character `public_id`, never by the
...
    bugs easier to exploit.

    ## Pagination

...

...

...

...

...

...

...

  contact:
...
servers:
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Create an account
      description: |
        Sends a verification email and returns a pending account.
        Rate limited to 5/hour, 20/day per subject.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, username, password, consent]
              properties:
                email: { type: string, format: email }
                username: { type: string, minLength: 3, maxLength: 32, example: mira }
                password: { type: string, minLength: 10, format: password }
                display_name: { type: string, maxLength: 64 }
                consent:
                  type: boolean
                  description: Acceptance of the terms and privacy policy. Required.
                marketing_consent: { type: boolean, default: false }
      responses:
        "201":
          description: Account created, pending email verification
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "409":
          description: Email or username already taken
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                duplicateEmail:
                  value:
                    ok: false
                    error:
                      code: email_taken
                      message: Этот адрес электронной почты уже зарегистрирован.
                      fields: { email: Этот адрес электронной почты уже зарегистрирован. }
        "422":
          description: Validation failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Confirm an email address
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, description: Single-use token from the email link. }
      responses:
        "200":
          description: Address confirmed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "422":
          description: Invalid, expired or already-used token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/resend-verification:
    post:
      tags: [Auth]
      summary: Resend the verification email
      description: Always returns the same response, so it cannot be used to discover which addresses are registered.
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { email: { type: string, format: email } }
      responses:
        "200": { description: Accepted (outcome deliberately not disclosed) }
  /auth/login:
    post:
      tags: [Auth]
      summary: Sign in
      description: |
        Accepts a username or an email address as the identifier.
        Every failure returns the identical message, so the endpoint cannot be
        used to enumerate accounts. Repeated failures lock the account
        temporarily.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier: { type: string, example: mira }
                password: { type: string, format: password }
      responses:
        "200":
          description: Signed in; sets `harmony_access`, `harmony_refresh`, `harmony_csrf`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401":
          description: Invalid credentials (identical for unknown user and wrong password)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                ok: false
                error: { code: invalid_credentials, message: Неверный логин или пароль. }
        "403":
          description: Account unverified, suspended or banned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: Too many attempts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate the session
      description: |
        Issues a new access token and rotates the refresh token. Replaying a
        rotated refresh token revokes **all** sessions for that account, which
        is the standard response to a suspected token theft.
      security: []
      responses:
        "200":
          description: Rotated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401":
          description: Missing, invalid or revoked refresh token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Sign out
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                all_sessions:
                  type: boolean
                  default: false
                  description: Revoke every session, not just this one.
      responses:
        "200":
          description: Signed out; auth cookies cleared
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /auth/me:
    get:
      tags: [Auth]
      summary: Current session
      responses:
        "200":
          description: The signed-in user
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      user: { $ref: "#/components/schemas/User" }
        "401": { description: Not authenticated }
...
  /auth/sessions/{jti}:
    delete:
      tags: [Auth]
      summary: Revoke one session
      parameters:
        - name: jti
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Revoked }
        "404": { description: Session not found }
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset
      description: Always returns the same response regardless of whether the address exists.
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { email: { type: string, format: email } }
      responses:
        "200": { description: Accepted }
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Complete a password reset
      description: The token is single-use and expires after one hour. All sessions are revoked.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 10, format: password }
      responses:
        "200": { description: Password changed }
        "422": { description: Invalid or expired token }
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Change the password
      description: Requires the current password. Revokes all sessions, including this one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password }
                new_password: { type: string, minLength: 10, format: password }
      responses:
        "200": { description: Password changed }
        "401": { description: Current password incorrect }
  /auth/reauthenticate:
    post:
      tags: [Auth]
      summary: Step-up authentication
      description: |
        Returns a short-lived token that authorises one sensitive action
        (such as deleting an account). A hijacked session alone is therefore
        not enough to destroy an account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties: { password: { type: string, format: password } }
      responses:
        "200":
          description: Step-up token
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      reauth_token: { type: string }
                      expires_in: { type: integer, example: 600 }
        "401": { description: Password incorrect }

  # --------------------------------------------------------------- Users ---
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: harmony_access
      description: |
        Access token set by `/auth/login`. Send `X-CSRF-Token` on unsafe methods.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: For non-browser clients. Exempt from CSRF.
...
  schemas:
    Ok:
      type: object
      properties:
        ok: { type: boolean, const: true }
        data: { type: object, additionalProperties: true }
    Error:
      type: object
      properties:
        ok: { type: boolean, const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            fields:
              type: object
              additionalProperties: { type: string }
              description: Field-level messages, so a form can highlight every problem at once.
            request_id: { type: string, description: Correlates with the server log. }
            retry_after: { type: integer }
            detail: {}
...
    AuthResponse:
      type: object
      properties:
        ok: { type: boolean }
        data:
          type: object
          properties:
            user: { $ref: "#/components/schemas/User" }
            expires_in: { type: integer }
            csrf_token: { type: string }
            requires_email_verification: { type: boolean }
            message: { type: string }
    User:
      type: object
      description: |
        The public projection. Never includes the email unless the target has
        opted in, nor any moderation state, counter or token.
      properties:
        public_id: { type: string }
        username: { type: string }
        display_name: { type: [string, "null"] }
        initials: { type: string }
        bio: { type: [string, "null"] }
        avatar_url: { type: [string, "null"] }
        avatar_color: { type: string }
        location: { type: [string, "null"] }
        website: { type: [string, "null"] }
        pronouns: { type: [string, "null"] }
        role: { type: string, enum: [member, moderator, admin] }
        is_verified: { type: boolean, description: Active account with a confirmed email. }
        posts_count: { type: integer }
        comments_count: { type: integer }
        followers_count: { type: integer }
        following_count: { type: integer }
        created_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
        email: { type: string, description: Present only if the target enables it. }
        profile_visibility: { type: string }
        is_self: { type: boolean }
...